polished more

+++ src/AbilityAction.h
@@ -10,9 +10,9 @@ namespace BOSS
     struct AbilityAction
     {
         ActionType  type;                // type of the ability
-        size_t      frameCast;           // the frame the ability was cast
-        size_t      targetID;            // index inside m_units of GameState class
-        size_t      targetProductionID;  // index inside m_units of the unit being produced by the target of this ability
+        int         frameCast;           // the frame the ability was cast
+        int         targetID;            // index inside m_units of GameState class
+        int         targetProductionID;  // index inside m_units of the unit being produced by the target of this ability
         ActionType  targetType;          // type of unit it was used on
 
 
@@ -23,7 +23,7 @@ namespace BOSS
           type = action;
           }*/
 
-        AbilityAction(ActionType newType, size_t newFrameCast, size_t newTargetID, size_t newTargetProductionID, ActionType newTargetType)
+        AbilityAction(ActionType newType, int newFrameCast, int newTargetID, int newTargetProductionID, ActionType newTargetType)
         {
             type = newType;
             frameCast = newFrameCast;
