+++ src/CombatSearch_BestResponse.h
@@ -16,11 +16,11 @@ namespace BOSS
 
     class CombatSearch_BestResponse : public CombatSearch
     {
-	virtual void                    recurse(const GameState & s, size_t depth);
+	virtual void recurse(const GameState & s, int depth);
 
-        CombatSearch_BestResponseData   m_bestResponseData;
+        CombatSearch_BestResponseData m_bestResponseData;
 
-        BuildOrderAbilities             m_bestBuildOrder;
+        BuildOrderAbilities m_bestBuildOrder;
 
     public:
