+++ src/Eval.h
@@ -11,8 +11,8 @@ namespace BOSS
     namespace Eval
     {
         //double ArmyCompletedResourceSum(const GameState & state);
-        double ArmyTotalResourceSum(const GameState & state);
-        double ArmyResourceSumToIndex(const GameState & state, int finishedUnitsIndex);
+        float ArmyTotalResourceSum(const GameState & state);
+        float ArmyResourceSumToIndex(const GameState & state, int finishedUnitsIndex);
 
         bool BuildOrderBetter(const BuildOrderAbilities & buildOrder, const BuildOrderAbilities & compareTo);
         bool StateBetter(const GameState & state, const GameState & compareTo);
